Move test-only SemanticDB code to test sources
